Fixed random logic at main.py
In previous version, when user has given their input and click the button, self.button.bind will choose between the two functions build2 or build3 randomly. However this randomizing process only happen once and after the first press occurred, the button will be bind to the one function it has already chose initially, therefore unable to re-choose the functions again randomly even if the user change their input. 

In this version, self.button.bind behavior is changed. when the button is pressed, it'll run a function that will select between the two build functions randomly instead of directly doing so in the self.button.bind.